refactor: move ToggleContainer to common component
