+++ src/components/common/ToggleContainer.js
@@ -1,0 +1,11 @@
+// import React from 'react';
+import { styled } from '@mui/material/styles';
+
+const ToggleContainer = styled('div')(() => ({
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'space-between',
+    width: '100%',
+  }));
+
+export default ToggleContainer;
