+++ src/components/main/ShowProgress.js
@@ -1,16 +1,9 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import { styled } from '@mui/material/styles';
 import Typography from '@mui/material/Typography';
 import FormControlLabel from '@mui/material/FormControlLabel';
 import { Button } from '@mui/material';
-
-const ToggleContainer = styled('div')(() => ({
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'space-between',
-  width: '100%',
-}));
+import ToggleContainer from '../common/ToggleContainer';
 
 // eslint-disable-next-line react/prop-types
 const ShowProgress = ({ setToggle, toggle, handleModalClose }) => {
